Multisite: Handling deactivation cases where the blog is still using the default settings
If SuperPWA is deactivated and if the blog is using the default settings (i.e, settings not saved to database yet).

SuperPWA ships with smart defaults. Some blogs might not change the default settings.
Default settings are not saved to the database until the user clicks the "Save Settings" in the plugin UI. (refer superpwa_get_settings() )
If that is the case, there is nothing more to clean up during uninstall, and we can remove the blog id from our list.

If settings are saved in the database, we want to preserve it, in case the user activates the plugin again.
These settings are removed during uninstall.